:bug: fix: correct  'res.code( )' to 'res.status( )'

+++ src/services/attorney.service.ts
@@ -5,6 +5,7 @@ import {
   deleteAttorneyModel,
   getAllAttorneyModel,
   getByIdAttorneyModel,
+  getByOabAttorneyModel,
   updateAttorneyModel,
 } from '../models/attorney.model'
 
@@ -12,28 +13,46 @@ export const createAttorneyService = async (
   req: FastifyRequest,
   res: FastifyReply
 ) => {
-  const shemaBody = z.object({
+  const schemaBody = z.object({
     nome: z.string().trim(),
     oabNumero: z.string().trim(),
   })
 
-  const { nome, oabNumero } = shemaBody.parse(req.body)
+  const body = schemaBody.safeParse(req.body)
+
+  if (!body.success) {
+    return res.status(400).send({
+      error: 'Dados inválidos',
+      issues: body.error.format(),
+    })
+  }
 
   try {
-    const advogado = await createAttorneyModel(nome, oabNumero)
+    const advogadoExist = await getByOabAttorneyModel(body.data.oabNumero)
 
-    return res.status(201).send(advogado)
+    if (advogadoExist) {
+      res.status(409).send({ error: 'Advogado ja existe no banco dea dados' })
+      return
+    }
+
+    const newAdvogado = await createAttorneyModel(
+      body.data.nome,
+      body.data.oabNumero //createAttorneyModel
+    )
+
+    return res.status(201).send({ message: 'Advogado criado com sucesso' })
   } catch (error) {
-    console.error('Erro ao criar advogado:', error)
-    return res.code(500).send({ error: 'Erro interno do servidor' })
+    console.error('Erro ao criar alocação:', error)
+    res.status(500).send({ error: 'Erro interno do servidor teste 01' })
   }
 }
+
 export const getAllAttorneyService = async (
   req: FastifyRequest,
   res: FastifyReply
 ) => {
-  const advogado = getAllAttorneyModel()
-  return res.send(advogado)
+  const advogado = await getAllAttorneyModel()
+  return res.status(200).send(advogado)
 }
 
 export const getByIdAttorneyService = async (
@@ -46,9 +65,17 @@ export const getByIdAttorneyService = async (
 
   const { id } = schemaParams.parse(req.params)
 
+  const advogadoExist = await getByIdAttorneyModel(id)
+
+  if (!advogadoExist) {
+    res.status(404).send({ error: 'Advogado não ja existe no banco dea dados' })
+    return
+  }
+
   const advogado = await getByIdAttorneyModel(id)
   res.status(200).send(advogado)
 }
+
 export const updateAttorneyService = async (
   req: FastifyRequest,
   res: FastifyReply
@@ -64,9 +91,17 @@ export const updateAttorneyService = async (
   const { id } = schemaParams.parse(req.params)
   const { nome, oabNumero } = schemaBody.parse(req.body)
 
+  const advogadoExist = await getByIdAttorneyModel(id)
+
+  if (!advogadoExist) {
+    res.status(404).send({ error: 'Advogado não ja existe no banco dea dados' })
+    return
+  }
+
   await updateAttorneyModel(id, nome, oabNumero)
   res.status(200).send({ message: 'Advogado atualizado com sucesso' })
 }
+
 export const deleteAttorneyService = async (
   req: FastifyRequest,
   res: FastifyReply
@@ -76,6 +111,13 @@ export const deleteAttorneyService = async (
   })
 
   const { id } = schemaParams.parse(req.params)
+
+  const advogadoExist = await getByIdAttorneyModel(id)
+
+  if (!advogadoExist) {
+    res.status(404).send({ error: 'Advogado não ja existe no banco dea dados' })
+    return
+  }
 
   await deleteAttorneyModel(id)
   res.status(200).send({ message: 'Advogado deletado com sucesso' })
